+++ CHANGELOG.md
@@ -93,6 +93,7 @@ keybinds
 - The borders of browser panes are nolonger affected by the scrollbar track being set
 - Adding a directory to the queue now orders its songs by `directories_sort` instead of relying on MPD's default order
 - `addyt` sometimes failing when files with identical file name already exists it the dir
+- Sixel image disappearing when switching tmux panes
 
 ## [0.11.0] - 2026-02-01